Rename Decimal() to Float64().

+++ pkg/models/validators.go
@@ -67,7 +67,7 @@ type RateIsValid struct {
 
 // IsValid adds an error if the value is not between 0 and 1.
 func (v *RateIsValid) IsValid(errors *validate.Errors) {
-	if v.Field.Decimal() < 0 || v.Field.Decimal() > 1 {
+	if v.Field.Float64() < 0 || v.Field.Float64() > 1 {
 		errors.Add(validators.GenerateKey(v.Name), fmt.Sprintf("%s must be between 0 and 1", v.Name))
 	}
 }
